pyproject.toml
[project]
name = "pythonproject"
version = "0.1.0"
description = "FastAPI access diagnosis service"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi==0.136.1",
    "httpx==0.28.1",
    "openai>=2.32.0",
    "pydantic-settings==2.14.0",
    "uvicorn==0.46.0",
]

[dependency-groups]
dev = [
    "pytest==9.0.2",
    "pytest-asyncio==1.3.0",
]

[tool.uv]
package = false

[tool.pytest.ini_options]
addopts = "-p no:cacheprovider"
asyncio_mode = "auto"
norecursedirs = [".*", ".venv", ".uv-cache", ".uv-python", ".uv-venv", "pytest-cache-files-*"]
pythonpath = ["."]
